refactor(kernel): extract observability syscalls to kernel/observability.rs
Phase G.3 of Phase 3 restructure plan.  Combines two related sections
into one observability submodule:

  - rust/kernel/src/kernel/observability.rs (new): impl Kernel block
    over the Rust observer registry (register_observer /
    unregister_observer / dispatch_observers / flush_observers /
    observer_count / dispatch_event), the file watch registry
    (register_watch / unregister_watch / match_watches), the
    sys_watch syscall, and the shared dispatch_mutation helper.
    ~135 LOC.
  - rust/kernel/src/kernel/mod.rs: drops the corresponding methods
    leaving "moved to" breadcrumbs at the section dividers; declares
    `mod observability;` alongside `mod locks;`.

dispatch_mutation, previously a private `fn`, becomes
`pub(super)` so the syscall implementations in mod.rs (sys_write /
sys_unlink / etc.) keep calling it via `self.dispatch_mutation(...)`.

The crate-root `use` import for `dispatch::*` drops the now-unused
`FileEvent` symbol; FileEventType and MutationObserver remain
because mod.rs still references them in struct definitions and
dispatch_mutation call sites.

cargo check -p kernel: clean.
